🎉 factorisation des méthodes 'crud' dans le BaseController, c'est super ! il y a presque rien a faire 🙂
