+++ aluilcine/src/main/java/fr/iut/aluilcine/controllers/UserController.java
@@ -22,79 +22,5 @@ import static org.springframework.http.HttpStatus.*;
 @RequestMapping("/users")
 public class UserController extends BaseController<User, UserRepository>{
 
-    @GetMapping("")
-    public ResponseEntity<List<User>> getAll() {
-        try {
-            return new ResponseEntity<>(repository.findAll(), OK);
-        }catch (Exception e){
-            customLogError(e.getMessage());
-        }
-        return new ResponseEntity<>(INTERNAL_SERVER_ERROR);
-    }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<User> getOneById(@PathVariable("id") String id){
-        try {
-            Optional<User> opt = repository.findById(id);
-            if (opt.isEmpty()){
-                return new ResponseEntity<>(NOT_FOUND);
-            }
-            return new ResponseEntity<>(opt.get(), OK);
-        }catch (Exception e){
-            customLogError(e.getMessage());
-        }
-        return new ResponseEntity<>(INTERNAL_SERVER_ERROR);
-    }
-
-    @PostMapping("")
-    public ResponseEntity<?> addOne(@RequestBody User user){
-        try {
-            // valide l'entite
-            Optional<String> optValidation = validateEntity(user);
-            if (optValidation.isPresent()) {
-                return new ResponseEntity<>(optValidation.get(), UNPROCESSABLE_ENTITY);
-            }
-            return new ResponseEntity<>(repository.save(user), CREATED);
-        }catch (Exception e){
-            customLogError(e.getMessage());
-        }
-        return new ResponseEntity<>(INTERNAL_SERVER_ERROR);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<?> updateOne(@PathVariable("id") String id, @RequestBody User user) {
-        try {
-            // valide l'entite
-            Optional<String> optValidation = validateEntity(user);
-            if (optValidation.isPresent()) {
-                return new ResponseEntity<>(optValidation.get(), UNPROCESSABLE_ENTITY);
-            }
-
-            Optional<User> opt = repository.findById(id);
-            if (opt.isEmpty()){
-                return new ResponseEntity<>(NOT_FOUND);
-            }
-
-            User _user = opt.get();
-            _user.setPseudo(user.getPseudo());
-            _user.setPassword(user.getPassword());
-            _user.setEmail(user.getEmail());
-
-            return new ResponseEntity<>(repository.save(_user), HttpStatus.OK);
-        }catch (Exception e){
-            customLogError(e.getMessage());
-        }
-        return new ResponseEntity<>(INTERNAL_SERVER_ERROR);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteOne(@PathVariable("id") String id) {
-        try {
-            repository.deleteById(id);
-            return new ResponseEntity<>(NO_CONTENT);
-        } catch (Exception e) {
-            customLogError(e.getMessage());
-        }
-        return new ResponseEntity<>(INTERNAL_SERVER_ERROR);
-    }
 }
